add service counterfactualsrv from old identity, add new identitysrv
